'Duke.java' renamed to 'DogeBot.java', refactored code to switch cases (for throwing exceptions later)
